+++ src/pages/HomePage.tsx
@@ -1,6 +1,6 @@
 import Contacts from "@components/Contacts";
 import Hero from "@components/Hero";
-import Projects from "@components/folder1/Projects";
+import Projects from "@components/projects/Projects";
 import { Fragment } from "react";
 
 const HomePage = () => {
